backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
# langchain-core 在 Python>=3.12.4 上要求 pydantic>=2.7.4
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pandas==2.2.1
dashscope==1.18.0
baidu-aip==4.16.10
python-dotenv==1.0.1
email-validator==2.1.1
pytest==8.1.1
httpx==0.27.0
langgraph==0.2.18
langchain-core>=0.2.20,<0.3.0
langchain-community>=0.2.12,<0.3.0
faiss-cpu>=1.7.4
